Fix split on space test for Perl 5.18.0
Change one test that now depends on the Perl version, and add another
to cover the old case.

In perldelta 5.18.0 it reads:

split's first argument is more consistently interpreted

After some changes earlier in v5.17, split's behavior has been
simplified: if the PATTERN argument evaluates to a string containing
one space, it is treated the way that a literal string containing one
space once was.